requirements.txt
einops==0.4.1
h5py==3.6.0
numpy==1.21.2
opencv_python==4.5.3.56
Pillow==9.2.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.2
